+++ taf/repository_tool.py
@@ -375,6 +375,19 @@ class Repository:
             if target_path.is_file():
                 self._add_target(targets_obj, str(target_path), previous_custom)
 
+    def delete_unregistered_target_files(self, targets_role="targets"):
+        """
+        Delete all target files not specified in targets.json
+        """
+        targets_obj = self._role_obj(targets_role)
+        for filepath in self.targets_path.rglob("*"):
+            if filepath.is_file():
+                file_rel_path = str(
+                    Path(filepath).relative_to(self.targets_path).as_posix()
+                )
+                if file_rel_path not in targets_obj.target_files:
+                    filepath.unlink()
+
     def _get_target_repositories(self):
         repositories_path = self.targets_path / "repositories.json"
         if repositories_path.exists():